rename NumericQuestionForm component, refactor NumericQuestionForm container
